tests: Clean up drive extension test setup
Replace make_env(), returning vm, dom, drives, with make_vm() returning
a vm ready to test. This is a preparation for replacing the context
manager with a pytest fixture.

We can get the drives from the vm using getDiskDevices() and access the
dom using vm._dom. Accessing the private _dom is ugly but this fine for
the test.

Change-Id: I41503b4c80717d7d83231d1bc205cd07a7759293
Signed-off-by: Nir Soffer <nsoffer@redhat.com>