tests: Remove empty TestReplication class
This class had a long comment about required replication tests, and no
tests. It looks like was a preparation for adding replication tests that
were never added.

The base class for the tests include a check function that seems to
check replication, so the comments about replication may be stale now.

Remove the empty class and keep the comments in the module docstring for
later inspection.

Now that we have only one test class, there is no need for the base
class, so merge both into the same class.

Change-Id: Ie4bc22149fe7612abb71017807a98fc3d322284e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>