snapshot: don't consider metadata without recovery
When we execute the snapshot we set details of the snapshot job into the
VM metadata. Unfortunetly, we read this metadata regarless the flow we
are in. That means, if we failed in executing a snapshot, we will get
the metadata from the VM when doing another snapshot.

This patch let us consume the snapshot metadata from the VM only on
recovery mode.

Change-Id: I88cda3a991f517acef238c2a68aee38e558f2bf9
Bug-Url: https://bugzilla.redhat.com/1984209
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>