numa: drop @cache.memoized and cache capabilities manually
The _numa() call is cached (memoized) based on arguments used when
calling the function. The _numa() function optionally took libvirt
capabilities as an argument. This argument however was never used,
except in tests. Meaning that _numa() was normally called only without
arguments in which case libvirt capabilities were evaluated only the
first time the function was called (by _numa() function itself).

Recent patch changed how we treat _numa() in Host.getCapabilities API
call. We fetch fresh libvirt capabilities and pass it as argument to
_numa(). This causes two problems. First, it now causes small leak
because the cache is allowed to grow indefinitely. Secondly, the results
of re-evaluated _numa() call are not available to the rest of the VDSM
code that calls _numa() without arguments (e.g. in sampling.py for every
sample).

The first problem could be easily solved by changing the memoizing
decorator to functools.lru_cache(maxsize=N). But to solve also also the
second problem the caching cannot be done based on arguments. We need
something special because we don't want to fetch the libvirt
capabilities every time we need to call _numa() and we also
don't want to re-evaluate _numa() on every call.

This patch removes the @cache.memoized decorator and creates caching in
numa.py There is a separate update call that fetches fresh libvirt
capabilities. The capabilities are examined only when they change.

Change-Id: I9f7bc5596fb3b3c25dcf585e3d3a07a3a1a86a9e
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>