virt: user proper disk type when changing CD
So far we always use FILE disk type when changing CD. This is wrong and
in some cases libvirt detect that and fails with

    libvirt.libvirtError: internal error: unable to execute QEMU command
    'blockdev-add': 'file' driver requires '/rhev/data-center/...' to be
    a regular file

Use appropriate disk type setup by prepareVolumePath() call.

Add tests for file and block based storage. Network disk doesn't need to
be handled in the code and tested as network disks are not supported as
CDROM devices.

Change-Id: Id96059357a191d665bfdbfc2e87e63babb4b19b4
Bug-Url: https://bugzilla.redhat.com/1990268
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>