tests: Use pytest.timeout instead of py-watch
py-watch has many issues:

1. Working in the test suite level, which needs endless maintenance when
   adding new tests and making the test suite slower.

2. Because of 1, it does not react quickly enough when a test deadlocks.
   We have to wait 10 minutes (20 in storage tests) until we time out.

3. It requires installing python debuginfo packages, which slows down
   test setup.

4. Even when debuginfo package are installed, we usually do not get a
   backtrace anyway.

5. This is another infrastructure we invented that nobody else is using,
   and we have to maintain it.

Instead of pywatch, use pytest-timeout plugin, maintained by the python
community. This works in the test level, so we can fail quickly when a
test deadlocks, and no maintenance is needed when adding new tests. We
always get a python backtrace when the test process is killed.

The disadvantage is that we don't get a backtrace of the underlying
python C code. But in the last 8 years we needed it only once or twice.
For these cases we can manually run specific tests via gdb.

py-watch is still use by the legacy nose tests, so we don't delete it
and the required debuginfo install step.

Change-Id: Iaa9327b76754f89892e3b0e4887f3c99f8c42ca4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>