tests: Improve failing tests
test_wrong_backing_file and test_unexpected_backing_file time out in the
CI in:

    File "tests/storage/hsm_test.py", line 136, in test_wrong_backingfile
       op.run()
    File "lib/vdsm/storage/operation.py", line 82, in run
       out, err = self._proc.communicate()
    File "/usr/lib64/python3.6/subprocess.py", line 863, in communicate
       stdout, stderr = self._communicate(input, endtime, timeout)
    File "/usr/lib64/python3.6/subprocess.py", line 1534, in _communicate
       ready = selector.select(timeout)
    File "/usr/lib64/python3.6/selectors.py", line 376, in select
       fd_event_list = self._poll.poll(timeout)

The tests time out when running qemuimg.rebase(unsafe=True). This is a
fast metadata operation that takes no time on a real system.

I don't see any reason for this timeout, but we try to rebase on an
empty file, and we don't specify the image and backing file format. In
this case qemu has to probe the image format and the backing file
format, and maybe there is a bug in qemu-img when rebasing on an empty
image.

Change the test to rebase on a non-empty raw image, and specify the
image and backing file format properly.

Change-Id: Iacf45617e89eafe1acdec40b48db2427d1a79d19
Signed-off-by: Nir Soffer <nsoffer@redhat.com>