tests: Always use backingFormat
When using "qemu-img create" or "qemu-img rebase" with a backing file,
we must always specify the backing format. Using backing file without
specifying the format is insecure and was deprecated long time ago. This
became a hard error in qemu-img 6.1.

In the real code we always specify the backing format, but in the tests
we skipped this in some places. Fix the tests and test infrastructure to
always specify the backing format.

Change-Id: Ib07860c09f61d96d2b6e8d3f054f034b4b7b922f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>