ci: Use el8 hosts
Lot of tests using qemu-img time out when running on Centos Stream 9
nodes. Ehud suggested to use host-distro: same to avoid these hosts, but
there are no el8stream hosts:
https://jenkins.ovirt.org/label/el8stream/

But we have el8 hosts:
https://jenkins.ovirt.org/label/el8/

Change host-distro to el8 to run on the available hosts.  We will
probably change this later to el8stream when we will have such hosts in
the CI.

Change-Id: Ia273d3a463982701da4ff03d0ca209eb870baf82
Signed-off-by: Nir Soffer <nsoffer@redhat.com>