snapshot: don't invoke abort to non-memory
The abort mechanism is mostly for the period we get stuck on the pivot
(libvirt snapshot call). Without memory, the operation don't take long
time to execute, as there is no memory dump. If we get into the abort
timeout with no memory snapshot, it means we got stuck on other
operations such as freezing or thawing. We can't abort those operations.
Therefore, the abort doesn't bring any benefit in the no memory snapshot
flow and removed for this flow.

Change-Id: I081855b761c798421f17e9d6e912b46e5e8f0f39
Bug-Url: https://bugzilla.redhat.com/1985973
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>