virt: add shutdown timer after console disconnect
This patch adds a delay before calling the shutdown() method
of a vm after user disconnects from a console if the
consoleDisconnectAction is set to 'Shutdown'. The duration of
this delay is set by the user in ovirt-engine webui and passed to
vdsm when a VM ticket is created.

VDSM will now wait for n minutes after user disconnects from console
and then start shutting down the VM by calling vm.shutdown(), the rest of
the process remains the same. The shutdown can be interrupted during this
period if a user reconnects to the console.
The vm.shutdown() method has a 'delay' argument which could theoretically
used in this case, however it serves a different purpose than what this
patch needs so it's better not to combine them.

Change-Id: Ib5be885fa78f6e53b778d6526ebae92e7897a7ee
Bug-Url: https://bugzilla.redhat.com/1944834
Signed-off-by: Filip Januska <fjanuska@redhat.com>