snapshot: introduce non-memory timeout
Although the abort thread is irrelevant for the non-memory snapshot
flow, we do have limitations. If the pivot did happen, we shouldn't
abort it. But, if it happens before, it's most likely to get stuck on
freeze operation. We can't abort the freeze and we get response only
when the file system is frozen. For Windows VMs there is a limitation
of 10 minutes file system freeze. With that, having a snapshot that
took more than 10 minutes to be frozen, will make the data inconsistent.
10 minutes will work fine if it's only one file system, so we can be
sure we are frozen, but we go on the safe side, which is 8 minutes to
get 2 minutes margin to execute the pivot and to have enough time to let
the whole file system to be frozen. The API will respect other given
value for the new timeout if override is desired.

Change-Id: Ia97b218e8566c8ca297355216a53739ae07683dc
Bug-Url: https://bugzilla.redhat.com/1985973
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>