osinfo: Remove unnecessary '.decode' calls
'rpm' library for a long time shipped a hack that added dummy
'decode' method on strings as a compatiblity layer for codebases
migrating from py2 to py3. This hack seems to be removed in the
latest versions, so we should adapt the code to the fact that
the returned values are already strings.

Change-Id: I824407d374b36f1e4444fea93a1f321008e45dd2
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>