net: Add retry to dumping of bond options
With newer kernel (>=4.18.0-338) there seems to be a race within
bonding driver. The bond created via sysfs hangs in state where
the device exists but the sysfs infrastructure for it is
missing:

ls: cannot access '/sys/devices/virtual/net/bond_test': No such file or directory
d??????????  ? ?    ?                               0 ?            ? bond_test

That causes the vdsm-tool dump-bonding-options to fail on
opening the sysfs bond mode file which should be always present.

To prevent that add retry that checks if the bond mode sysfs
file is valid, if not remove the bond and create it again.
This makes the driver to be responsive again.

Change-Id: I41d06ba1dd9fc8de113b7fc57891ff90598f7372
Bug-Url: https://bugzilla.redhat.com/1998865
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit 802caa1b6bfbd4f08933a48d739883ef11c82545)