storage: don't update device mapper size on SPM during LV extend
When the disk is extended during VM migration, data corruption would
happen if the destination host doesn't update the size of the disk (see
BZ #1883399). To avoid this situation, we refresh the disk on the
destination host first to ensure disk size is correct. If the
destination host doesn't support disk refresh, we stop the disk monitor
which should prevent any further disk extension and also abort disk
refresh on the source host. VM eventually will be paused due
to running out of the disk space.

However, this is not complete solution and the data corruption can
happen in this specific case: when the destination host doesn't
support disk refresh, source host is SPM and disk is extended only
once. Disk extension happens on SPM and lvm updates device mapper when
running LV extend command, so SPM actually sees the extension even if we
don't run LV refresh on the host. Disk monitoring is disabled, so no
further extensions will happen, but the first one already has happened
and is visible by the source host (SPM). If VM writes to this extended
area until migration finishes, the disk corruption will happen, as this
area is not visible on the destination host.

To prevent this, add additional argument "--driverloaded n" to lvm
extend command. This will cause that the extension will not be visible
to the system (SPM host) and thus cannot be used by the VM.
The extension will be visible only when we refresh the disks from vdsm,
as this was an ordinary host.

Change-Id: Ic9ba4d41f79ebd95b00745c2428f7fd797e67113
Bug-Url: https://bugzilla.redhat.com/1983882
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>