lvm: use run_command() in removeLVs() flow
Modify removeLVs() flow to use new run_command() which now raises
LVMCommandError.

The exceptions raised in this flow now inherit from LVMCommandError and
provide more details for better debugging.

The origial error (CannotRemoveLogicalVolume) is used as a wrapper for
other errors so we can not change it to inherit from LVMCommandError
which is what we need in removeLVs(). In this case we can add a new
exception for this flow - LogicalVolumeRemoveError.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1536880
Change-Id: Ibca33d32dbdddc5cfff914a85807f5c688670592