virt: store CPU policy and pinned CPUs in metadata
This is a first patch of the CPU policies series. We should get the CPU
policy in VM metadata. If it is not there it has to be detected and
stored there so that we know it on recovery. Otherwise, when we start
managing the vCPUs and add pinning it may be falsely mistaken for manual
CPU pinning. For the same reason we also store which vCPUs are manually
pinned. Later, the vCPUs without pinning will be using pCPUs from shared
pool and we need to remember which vCPUs were pinned by the user and
which have pinning defined by VDSM.

Defined policy names are:

- none: no policy defined, CPUs from shared pool will be used
- pin: manual CPU pinning or NUMA auto-pinning policy
- dedicated: each vCPU is pinned to single pCPU that cannot be used by
  any other VM
- siblings: like dedicated but physical cores used by the VM are blocked
  from use by other VMs
- isolate-threads: like siblings but only one vCPU can be assigned to
  each physical core

Related feature page for policies is:
https://ovirt.org/develop/release-management/features/virt/dedicated-cpu.html

Change-Id: I7bc7bad82a20d47d06135c82fa58572a2327badd
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1782077