virt: migration support for dedicated CPUs
When migrating the VM we need to remove any CPU pinning that was defined
by VDSM. The reason is that pinned CPUs may serve a different purpose on
the destination host (may be dedicated to another VM) or it may not be
present on the destination at all.

For VMs with no policy the pinning is simply dropped and it will be
filled on destination. For VMs with manual CPU pinning or VMs that use
NUMA auto pinning this affects only vCPUs that don't have any pinning
defined by user and are using shared CPU pool. Such configuration is
also removed and it will be filled again on destination.

For VMs with a policy we expect Engine to pass a new pinning
configuration to the source VDSM. The information is passed in "cpusets"
parameter in the form of a list. Each item of the list corresponds to
vCPU and contains a string with cpuset definition.

Change-Id: I3de2d50e8ab26a8728beb662339fdbecb8aacf74
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1782077