virt: shared pool creation
NUMA code is extended to build a list of CPUs on a core for each core
available. This is then used to build a list of all CPUs in a shared
pool -- i.e. CPUs available to all VMs without any specific CPU policy.

The list is built in a way that:

- CPUs of VMs with manual CPU pinning or NUMA auto-pinning policy are
  included in shared pool

- CPUs of VMs with dedicated policy are excluded from shared pool

- CPUs of VMs with isolate-threads or siblings policy are excluded from
  shared pool and all their siblings as well, so that whole cores are
  removed from shared pool and left exclusive to the particular VM

The updates need to be exclusive and cannot run concurrently, otherwise
the assigned CPU sets maybe wrong. Two racing VM.destroy calls are
problematic because we could fail to increase the shared pool so the
configured CPU set will be smaller than it could be for some VMs. Racing
VM.destroy and VM.create is much more problematic as it can lead to
situations where dedicated CPU would be used by other VMs.

Change-Id: Ife3797cda4419ecd153a136dea9fe35663f07f18
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1782077