virt: update shared pool on VM start/stop
To keep the list of pCPUs in shared pool accurate it has to be updated on
every start and stop of a VM. More concretely:

When starting a VM without CPU policy its vCPUs need to be configured to
use pCPUs from shared pool.

When starting or stopping a VM with CPU policy the shared pool has to be
updated. Starting from a set of all physical CPUs, CPUs of a VM with
policy dedicated, isolate-threads or siblings are removed from a set.
For VMs with isolate-threads and siblings policies all siblings of all
their CPUs are also removed to make sure that whole cores are allocated.
Finally all VMs without policy need to have their vCPUs reconfigured to
use only CPUs from the newly constructed shared pool.

Change-Id: Id8a0e96d53971c9180da04498c0eeefdab6cdbe3
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1782077