el9stream: tests: lib: Adapt to changed 'SC_NPROCESSORS_ONLN' behaviour
'TestExecCmdAffinity' tests are here because we want to be sure that
even though we pin the vdsm process to a single CPU, other spawned
commands should run on all available CPUs by default.

The test functions are decorated with '@forked' so we can use 'taskset'
on the process executing the test itself without affecting other tests.
In 'testResetAffinityWhenConfigured' first we pin the test process
to a single CPU and then execute a separate 'sleep' process to see if it
runs on all CPUs.

The problem is the comparison:

 self.assertEqual(taskset.get(proc.pid), online_cpus())

which _dynamically_ retrieves the number of the active processors
for the current process with:

 os.sysconf('SC_NPROCESSORS_ONLN')

If we just pinned our process to a single CPU, why would we expect it to
return all the available CPUs? Weird, but it seems it indeed worked that
way... until el9stream, where the behavior is changed (and probably more
reasonable).

One can test that with interactive python console on both el8 and
el9stream:

 Python 3.x.x ...
   >>> import os
   >>> os.getpid()
   129348
   >>> os.sysconf('SC_NPROCESSORS_ONLN')
   8

then in a separate console limit the process's CPUs:

 $ taskset -c -p 0 129348
 pid 129348's current affinity list: 0-7
 pid 129348's new affinity list: 0

and try in the python console again:

   >>> os.sysconf('SC_NPROCESSORS_ONLN')

which will print out '8' in el8 and '1' in el9stream.

The fix in our case seems simple - let's save the real CPU
count before forking so we get consistent results.

Change-Id: I7d2116ab3c4503fd2e523f4314e8e55d83b13013
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>