Pylint: fix E1120 (no-value-for-parameter)
Pylint version update[1] revealed an E1120 (no-value-for-parameter)-
the exception constructor did not receive all required parameters.
We don't really reach the point where we run an LVM command, and the
exception seems to describe a scenario when incorrect tags are passed.
Therefore, the 'ValueError' exception seems logically appropriate.
Line 1409 contains a similar issue which is handled in another path[2].

[1] https://gerrit.ovirt.org/#/c/vdsm/+/116871/
[2] https://gerrit.ovirt.org/#/c/vdsm/+/116258

Signed-off-by: hbraha <hbraha@redhat.com>
Change-Id: I79609b5196a3177673b648c3965f2c15d9b2bdcf