lvm: Fix invalid argument handling
If lvm.changeLVsTags() is called with conflicting tags, we raised
se.LogicalVolumeReplaceTagError which raises a TypeError:

    TypeError: __init__() missing 3 required positional arguments: 'rc',
    'out', and 'err'

adding unwanted noise, and making it harder to the debug the real issue.
This was introduced in commit 14a89f65ccb6089f67d9dae465801ec1bf7db980

    lvm: use run_command() in changelv() flow

Changing se.LogicalVolumeReplaceTagError to inherit from
se.LVMCommandError.

We don't really reach the point where we run an LVM command, and the
exception seems to describe a scenario when incorrect tags are passed.
Therefore, the 'ValueError' exception seems logically appropriate.

Signed-off-by: hbraha <hbraha@redhat.com>
Change-Id: I79609b5196a3177673b648c3965f2c15d9b2bdcf