el9s: spec: Require 'crontabs'
We have an hourly cron job distributed with vdsm [1], so we should
require 'crontabs' package.

This was found with 'rpmlint' in el9s, reported as
'missing-dependency-to-cron'. The change in the spec doesn't fix the
'rpmlint' run though, as the check in the shipped version  (1.11 at the
moment of writing) is broken.  It requires a dependency to 'cron'
package, but the usual name for the package providing cron functionality
is 'cronie'. Furthermore, the dependency should actually be on
'crontabs', which is fixed in newer versions of 'rpmlint' [2].

[1] https://github.com/oVirt/vdsm/blob/b4bbb2df896d5ed15f4bc0c5ca4adff0335535e4/vdsm.spec.in#L870
[2] https://github.com/rpm-software-management/rpmlint/pull/404

Change-Id: Idf2b537e9ec0d6424319aeb3db78feacc12f9b94
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>