virt: get real vCPU count instead of maximum
The element value does not have to correspond to the real vCPU count. It
is a maximum number of vCPUs a VM can have, but the actual count may be
lower to make CPU hot-plugging possible. The actual count is specified
in the (optional) attribute "current". Only if the attribute is not
present the element value is the real count.

The original implementation returned None in case there is was no <vcpu>
element present. But missing <vcpu> suggests a broken domain XML so now
we raise an exception instead.

Change-Id: I67835dc2b05cb11cb7c0fcb6d0ce1c802f968c28
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>