spec: bump libvirt version
Bump libvirt version to get fix for BZ #2003644. This issue causes
CD change failure when original CD has defined startupPolicy and we
try to switch to block-based CD, where startupPolicy is not supported.

As required libvirt version (7.6.0-4) is not available on CentOS stream
yet, require this version only on RHEL for now.

Change-Id: I7461b05b2ae10582013f9d3986a7d311e47c0b0a
Bug-Url: https://bugzilla.redhat.com/1990268
Related-To: https://bugzilla.redhat.com/2003644
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>