snapshot: fix the recovery mechanism
The recovery mechanism was considered to be used only when the memory
operation is executed as this might take long time and something might
happen to VDSM. This might cause undesired behavior when we recover
after non-memory snapshot, which executes fast in libvirt but might take
time to freeze, or even if we missed libvirt execution of the snapshot,
meaning we lost VDSM for the exact period of time of the execution.

In such cases the snapshot might incorrectly finish successfully:
- If we are in the middle of the domain job, we will catch it during the
  recovery process since we check the domain job and report correctly.
- If we are after the domain job, it will act as succeeded(not precise,
  since the operation might have failed and we missed it).
- If we are before the domain job, it will act wrong, marking it as
  succeeded.

In this patch we determine if the snapshot succeeded by checking the
VM's domain XML. In it, we look for new volumes that were made for
the snapshot job.

Change-Id: I6c987f426d7333f866532fe5fa5aa215576cd833
Bug-Url: https://bugzilla.redhat.com/2012832
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>