supervdsm_server: Fix race when runAs times out
When waiting for runAs child process, if the process disappears right
after we checked if the process is alive, os.kill() can fail with
ProcessLookupError, hiding the Timeout error from the caller.

Fix by handling ProcessLookupError around os.kill(). Now handle the
error and and raise the expected Timeout exception in the caller.

Change the way we check if the process is alive according to
multiprocessing docs[1]. After calling join(1), we can check the
process exitcode instead of calling is_alive().

[1] https://docs.python.org/3/library/multiprocessing.html#multiprocessing.Process.join

Change-Id: I02dd4aa543aabfc582767ef932e87acb759351a2
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>