drivemonitor: Fix extend during backup
During backup, libvirt does not report the allocation for chunked drive.
If we enable debug logs, we will see:

    2021-10-18 02:12:48,025+0300 DEBUG (periodic/2) [virt.vm]
    (vmId='6e95d38f-d9b8-4955-878c-da6d631d0ab2') Extension info for drive sdb
    volume 624b1158-f2b9-4b4f-91c4-332e5594c8fb: BlockInfo(capacity=4294967296,
    allocation=0, physical=3221225472) (vm:1277)

We have the same issue before the guest writes to the drive, so
DriveMonitor.should_extend_drive() does not complain and returns False
for this case.

The result is that during backup, when we get a block threshold event,
or when the VM paused, we never extend the disk, so the VM cannot be
resumed until the backup is stopped.

Once backup is stopped, libvirt reports again the actual allocation, and
the vdsm extends the disk and resume the VM.

This issue is likely a bug in libvirt or qemu, but we can handle the
missing allocation in a smarter way. If we don't have allocation info,
but the drive block threshold was exceeded, we know that the guest wrote
to the disk, and we can assume that the disk needs extension.

Since this is not an excepted situation, we log new warning:

    2021-10-18 02:12:48,025+0300 WARN  (periodic/2) [virt.vm]
    (vmId='6e95d38f-d9b8-4955-878c-da6d631d0ab2') No allocation info for drive
    sdb, but block threshold was exceeded - assuming that drive needs extension
    (drivemonitor:253)

We may silence this warning later when we are more confident about this
situation, and after we resolve this issue with libvirt and qemu folks.

Since this is urgent fix that we want to get in 4.4.9, a test for this
scenario will be added later.

Change-Id: I73a22ef2187c2c88a622207b930cfdbcd6e7cb3c
Bug-Url: https://bugzilla.redhat.com/2015121
Signed-off-by: Nir Soffer <nsoffer@redhat.com>