livemerge: recover from failed pivot attempt
During live merge we call syncVolumeChain multiple times to make sure
the actual/libvirt chain is synced to current/vdsm chain.

When pivot starts, the new requested chain is passed to syncVolumeChain
which compares it to current vdsm chain. This way we can tell what
volume is being removed.

If the volume being removed is a leaf/active layer, it is flagged as
ILLEGAL in vdsm to prevent usage.

Then libvirt blockjob (abort) is started and if it fails the old code
never recovered the volume from ILLEGAL state and manual intervention
was required.

This patch adds a helper to switch the leaf volume back to LEGAL and
calls the helper if libvirt fails abort block job.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1949475
Change-Id: Ia57c26529cf60d381d3df143a4f7195948ce1cea