tests: add dedicated option for login into all tagets
Add an dedicated option for login into all tagets method. This is not
current vdsm flow and also has its drawbacks, but probably can be used
as an option to consider if cuncurrent login will turn out as not a
useable/safe way how to log in into multiple targets.

All add a docstring for this method, describing its drawbacks.

Change-Id: Ifca6a4bcb2ed30f0bfbe8fad0e962407df6f68d0
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>