drivemonitor: Fetch block info using block stats API
When extending volumes, we use libvirt.virDomain.blockInfo() to get
volume allocation. This API is easy to use, but it does not work for
backup scratch disks, volumes in the backing chain, or blockCopy
destination volume.

We want to replace usage of libvirt.virDomain.blockInfo() with
libvirt.virConnect.domainListGetStats(), that works for all block nodes,
include backup scratch disks[1] (since RHEL 8.6).

Vdsm already collects block stats for sampling purpose, but we cannot
used this code::

- We need the allocation info at the time of the call, and sampling
  collects values only every 15 seconds.
- Sampling does not collect info for the backing chain (and should not)
  but we must collect info for the backing chain.
- Sampling collects all stats while we need only block stats.
- Sampling collects stats for all VMs while we need only single VM.
- Sampling skip non responsive VMs, while we don't skip blockInfo()
  calls.
- Drive monitoring cannot depend on sampling, a subsystem with different
  requirements (best effort) and maintained by different teams.

Using libvirt.virConnect.domainListGetStats() is tricky, since it
requires a libvirt.virDomain object as parameter, and this object is
wrapped by the VM._dom object. Since the VM owns the _dom object, it is
natural to provide a method to get block stats in the VM object:
VM.get_block_stats().

Another problem with libvirt.virConnect.domainListGetStats() is the
unhelpful return value, flat mapping of "block.N.KEY" to VALUE for all
block nodes:

    {
        ...
        "block.0.fl.times": 0,
        "block.1.name": "sda",
        "block.1.path": "/rhev/.../44d498a1-54a5-4371-8eda-02d839d7c840",
        "block.1.backingIndex": 2,
        "block.1.rd.reqs": 13448,
        "block.1.rd.bytes": 415614976,
        "block.1.rd.times": 9940902315,
        "block.1.wr.reqs": 4909,
        "block.1.wr.bytes": 82999296,
        "block.1.wr.times": 47469574949,
        "block.1.fl.reqs": 683,
        "block.1.fl.times": 4204366339,
        "block.1.allocation": 216006656,
        "block.1.capacity": 6442450944,
        "block.1.physical": 7113539584,
        "block.1.threshold": 6576668672,
        "block.2.name": "sda",
        ...

There is lot of unneeded information for monitoring context, and no
easy way to extract the single value we actually need. A new method
DriveMonitor.get_block_stats() added, extracting this info in a useable
form:

    {
        2: drivemonitor.BlockInfo(
            index=2,
            name='sda',
            path='/rhev/.../44d498a1-54a5-4371-8eda-02d839d7c840',
            allocation=216006656,
            capacity=6442450944,
            physical=7113539584,
            threshold=6576668672,
        ),
        ...
    }

We will use this in extend flows to fetch block info for all volumes
when we try to extend drives.

[1] https://bugzilla.redhat.com/2017928

Change-Id: I8cdaaaf56c9f1e078809e4400a86219fc8086c41
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>