vm: Pass BlockInfo instead of unpacking it
Vm.getExtendInfo returns a BlockInfo named tuple, but we unpacked it and
sent multiple arguments. When we pass the BlockInfo to the drive
montior, we send it instead of unpacking it. This simplifies the code
and include more context when we log the named tuple.

This will be even more useful when BlockInfo will return more
information about the block node like backing index, drive name, and
path.

Change-Id: I3131b630b7d269c076917a7188442214159c0562
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>