virt: CPU hotplugging support for dedicated CPUs
When adding more CPUs for VMs with CPU policy engine has to dedicate the
new CPUs to match the policy. For this the API has to be extended to
allow passing the CPU sets for new VMs. Engine will decide whether or
not it can pass the argument based on cluster version. For older cluster
versions this argument does not have to be specified even on new VDSM
because VMs with CPU policy cannot be used there.

To make the API more flexible, we expect that engine passes
configuration for all CPUs and not just the new ones. This potentially
allows engine to relocate also the already assigned CPUs of the VM to
optimize use of resources on the host. When decreasing the number of
CPUs there is nothing special to do for the VM with dedicated CPUs.

In all cases the shared CPU pool has to be updated and VMs without
dedicated CPUs reconfigured.

Change-Id: I3f05f4ae71101513df0c34f17245fdad290f9e20
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1782077