net: Fix source route definition when missing
Source route that was deleted or went somehow missing on
the host could lead to unfixable out-of-sync status in engine.
The code for source routes tried to optimize when we generate
new source route state by checking if the gateway has changed between
configuration. The consequence is that if the gateway did not
change, the source route wouldn't be configured again. To prevent
that stop checking if the gateway has changed between configs. This
approach has been used for default route networks already.

Change-Id: I4f3cafecde5d0fb564c4ba8cdd49be8808cf8113
Bug-Url: https://bugzilla.redhat.com/2022354
Signed-off-by: Ales Musil <amusil@redhat.com>