livemerge: Do not use getExtendInfo for base volume
When extending base drive before live merge we used Vm.getExtendInfo()
to get the current block info. This is wrong for several reasons:

- The function returns block info for the drive active volume, instead
  of the base volume for the merge, which is never the active volume.

- The function will replace Drive.blockinfo with the new block info,
  which is unwanted side affect when we try to extend the base volume.

- It calls libvirt for no reason.

Since we need the volume capacity, add the capacity to job.extend dict
when starting a merge, so we don't need to get it from the storage API
on each call to _start_extend().

Change-Id: Ic4b6dcccab4336e72a000b511d0e10c003eac9c5
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>