storage: move connectStorageOverIser() function into IscsiConnection
Encapsulating connectStorageOverIser() into IscsiConnection object
allows us to simplify this method little bit and also unifies connection
of all storage server. Now, all storage servers are connected by calling
connect() method on respective object, without any need to call some
pre-connect method, which is specific to given storage server.

To check if initiatorName was passed into as a connection parameter
from the engine (which shouldn't be possible anyway as it's not covered
in IscsiConnectionParameters [1]) we need to create iface object of the
connection. However, the implementation of iscsi.IscsiInterface seems to
be buggy and throws KeyError when initiatorName is not set so we need to
check for KeyError when assing it. This would deserver fix and whole
module revision, but this is different task than BZ #1787192 and can be
done later.

[1] https://github.com/oVirt/vdsm/blob/v4.50.0/lib/vdsm/api/vdsm-api.yml#L379

Change-Id: I1549393cc75faf2d195d62c13f89491f07301d14
Bug-Url: https://bugzilla.redhat.com/1787192
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>