concurrent: Add Timer class
This patch adds a Timer class to the concurrent module.
The class is based on and behaves pretty much the same
as the threading.Timer class, except that the thread
which carries out the target function is created with
concurrent.thread instead of the regular
threading.Thread. This makes the Timer class available
for use in vdsm, while still ensuring every thread is
created with concurrent.thread.

The Timer also doesn't inherit from the threading.Thread
class, like the threading.Timer does, but instead keeps
the thread object as an attribute.

Change-Id: I28f7f0a7f254088129964bc7d30e5fae846eb3fb
Signed-off-by: Filip Januska <fjanuska@redhat.com>