vm: Replace blockInfo with block stats
Replace the call to blockInfo() with DriveMonitor.get_block_stats(),
returning block info for all volumes. To extract the block stats for the
drive active volume we need the index of the drive.

We get block stats only if we have drives that should be extended, or
when we pre-extend a drive when starting replication.

getExtendInfo() was modified to amend block info from libvirt with
information from the replica in case the drive is not chunked but
replicating to a chunked replica. This method should be removed once we
start using libvirt block stats for the replica drive.

Change-Id: I5600bedf886be993233df67dcad39078e7c920c8
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>